Extract composer site settings UI
