+++ assets/js/composer-post-commit-state.js
@@ -1,4 +1,4 @@
-import { parseEncryptedMarkdownEnvelope } from './encrypted-content.js?v=press-system-v3.4.27';
+import { parseEncryptedMarkdownEnvelope } from './encrypted-content.js?v=press-system-v3.4.28';
 
 export function createPostCommitStateApplier({
   stagingRegistry,
