+++ assets/press-system.json
@@ -1,13 +1,13 @@
 {
   "schemaVersion": 1,
   "type": "press-system",
-  "version": "3.4.27",
-  "tag": "v3.4.27",
+  "version": "3.4.28",
+  "tag": "v3.4.28",
   "upgradeFrom": {
     "ranges": [
-      ">=3.4.26 <3.4.27"
+      ">=3.4.27 <3.4.28"
     ],
     "allowUnknownSource": false,
-    "message": "Update to v3.4.26 first."
+    "message": "Update to v3.4.27 first."
   }
 }
