+++ scripts/test-system-release-package.sh
@@ -91,6 +91,11 @@ if ! grep -qx "press-system-${version}/assets/js/composer-index-tabs-ui.js" "${e
   exit 1
 fi
 
+if ! grep -qx "press-system-${version}/assets/js/composer-site-settings-ui.js" "${entries_file}"; then
+  echo "expected package to include composer Site Settings UI code" >&2
+  exit 1
+fi
+
 if ! grep -qx "press-system-${version}/assets/js/editor-file-tree-ui.js" "${entries_file}"; then
   echo "expected package to include editor file tree UI code" >&2
   exit 1
